python/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
fastapi==0.111.0
uvicorn==0.30.1
aiofiles==24.1.0
pydantic==2.7.4
